fix(build): expose new API
