+++ packages/ducks/src/bind-selectors/index.ts
@@ -1,2 +1,2 @@
-export * from './bind-selectors';
-export * from './connect-selectors-to-store';
+export { bindSelectors } from './bind-selectors';
+export { connectSelectorsToStore } from './connect-selectors-to-store';
