+++ packages/ducks/src/get-actions/action-creator-candidates.ts
@@ -1,0 +1,17 @@
+import { TypedAction } from '@ngrx/store/src/models';
+import {
+  ActionLoaded,
+  ActionPlain,
+  DispatchLoaded,
+  DispatchPlain
+} from '../create-duck/create-duck';
+
+export type ActionCreatorCandidates<TClass> = {
+  [TMember in keyof TClass]: TClass[TMember] extends TypedAction<infer TType> &
+    DispatchPlain
+    ? ActionPlain<TType>
+    : TClass[TMember] extends DispatchLoaded<infer TPayload> &
+        TypedAction<infer TType>
+    ? ActionLoaded<TType, TPayload>
+    : never
+};
