+++ packages/ducks/src/get-actions/action-creator-filter.ts
@@ -1,0 +1,5 @@
+import { ActionCreator } from '@ngrx/store';
+
+export type ActionCreatorFilter<T> = ({
+  [P in keyof T]: T[P] extends ActionCreator ? P : never
+})[keyof T];
