+++ packages/ducks/src/get-actions/action-creators.ts
@@ -1,0 +1,7 @@
+import { ActionCreatorCandidates } from './action-creator-candidates';
+import { ActionCreatorFilter } from './action-creator-filter';
+import { Constructable } from './constructable';
+export type ActionCreators<T extends Constructable> = Pick<
+  ActionCreatorCandidates<InstanceType<T>>,
+  ActionCreatorFilter<InstanceType<T>>
+>;
