+++ packages/ducks/src/get-actions/constructable.ts
@@ -1,0 +1,1 @@
+export type Constructable = new (...args: any) => any;
