+++ packages/ducks/src/public_api.ts
@@ -1,11 +1,16 @@
+export { bindSelectors } from './bind-selectors/bind-selectors';
+export { createDuck, dispatch } from './create-duck/create-duck';
 export * from './decorators/action.decorator';
 export * from './decorators/ducksify.decorator';
 export * from './decorators/initial-state.decorator';
 export * from './ducks/bind-selector-group';
 export * from './ducks/ducksify';
 export * from './ducks/effect';
+export { getActions } from './get-actions/get-actions';
+export { getReducer } from './get-reducer/get-reducer';
 export * from './operators/where-type';
-export * from './public_api_new';
 export * from './reducer/reducer-from';
+export { StoreFacade } from './store-facade/store-facade';
 export * from './typings/duck/duck';
 export * from './typings/duck/ducksified-action';
+export { usePick } from './use-pick/use-pick';
