--- packages/ducks/src/public_api_new.ts
@@ -1,6 +1,0 @@
-export * from './bind-selectors';
-export * from './create-duck';
-export * from './get-actions';
-export * from './get-reducer';
-export * from './store-facade';
-export * from './use-pick';
